requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
Django==4.2
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
importlib_resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
uritemplate==4.1.1
